feat: 검증 된 Card 모델 생성
