+++ README.md
@@ -10,11 +10,11 @@
 ---
 - 카드
   - 카드 목록은 ACE부터 KING까지 총 13장으로 구성되어 있다.
-    - [ ] A(1,11), J(10), Q(10), K(10)는 각각 괄호의 숫자로 반환된다.
-    - [ ] 그 이외의 숫자는 그 숫자로 반환된다.
+    - [x] A(1,11), J(10), Q(10), K(10)는 각각 괄호의 숫자로 반환된다.
+    - [x] 그 이외의 숫자는 그 숫자로 반환된다.
   - 각 카드는 4가지 문양을 가지고 있다.
-    - [ ] 스페이드, 다이아, 하트, 클로버로 구성된다. 
-  - [ ] 카드는 카드 문양과 숫자의 조합을 반환한다.
+    - [x] 스페이드, 다이아, 하트, 클로버로 구성된다. 
+  - [x] 카드는 카드 문양과 숫자의 조합을 반환한다.
 - 카드 더미
   - 카드 더미는 총 52장의 카드로 구성되어 있다.
     - [ ] ACE 부터 KING까지 13개의 숫자와 4가지 문양을 한 더미로 가지고 있다.
